PIG-861: POJoinPackage lose tuple in large dataset
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/tags/release-0.3.0@795936 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/builtin/BinStorage.java
@@ -52,9 +52,9 @@ import org.apache.pig.impl.util.WrappedIOException;
 
 
 public class BinStorage implements ReversibleLoadStoreFunc {
-    public static final byte RECORD_1 = 0x01;
-    public static final byte RECORD_2 = 0x02;
-    public static final byte RECORD_3 = 0x03;
+    public static final int RECORD_1 = 0x01;
+    public static final int RECORD_2 = 0x02;
+    public static final int RECORD_3 = 0x03;
 
     Iterator<Tuple>     i              = null;
     protected BufferedPositionedInputStream in = null;
@@ -70,7 +70,7 @@ public class BinStorage implements ReversibleLoadStoreFunc {
 
     public Tuple getNext() throws IOException {
         
-        byte b = 0;
+        int b = 0;
 //      skip to next record
         while (true) {
             if (in == null || in.getPosition() >=end) {
@@ -82,23 +82,23 @@ public class BinStorage implements ReversibleLoadStoreFunc {
             // After reading the second RECORD_1 in the above
             // sequence, we should not look for RECORD_1 again
             if(b != RECORD_1) {
-                b = (byte) in.read();
+                b = in.read();
                 if(b != RECORD_1 && b != -1) {
                     continue;
                 }
                 if(b == -1) return null;
             }
-            b = (byte) in.read();
+            b = in.read();
             if(b != RECORD_2 && b != -1) {
                 continue;
             }
             if(b == -1) return null;
-            b = (byte) in.read();
+            b = in.read();
             if(b != RECORD_3 && b != -1) {
                 continue;
             }
             if(b == -1) return null;
-            b = (byte) in.read();
+            b = in.read();
             if(b != DataType.TUPLE && b != -1) {
                 continue;
             }
